Optimize decode attention: batch decompress, bmm, native dtype
Five decode-path optimizations that together yield +8.7% (short context)
and +25.7% (medium context) throughput on RTX 5080 with Qwen3.5-35B-A3B:

- Batch value decompression across all blocks into a single matmul
- Replace per-KV-head Python loop with batched torch.bmm
- Use model-native dtype (bf16) instead of fp32 upcasts for GEMMs
- Store QJL signs as fp16 to skip int8->float cast on every decode step
- Incremental block compression: only compress newly appended tokens

+++ src/turboquant/vllm_plugin/compressor.py
@@ -84,7 +84,7 @@ class TQKeyCompressorGPU:
         residual = flat - k_mse
         r_norm = torch.norm(residual, dim=-1)  # (N,)
         projected = residual @ self.S.T  # (N, D)
-        signs = (projected >= 0).to(torch.int8) * 2 - 1  # {-1, +1}
+        signs = (projected >= 0).to(torch.float16) * 2 - 1  # {-1, +1} as fp16
 
         return {
             "k_mse": k_mse.half(),
